requirements.txt
jupyterlab==3.6.7
pandas==1.3.5
typing_extensions>=4.5
uvicorn==0.22.0
fastapi==0.103.2
pydantic==2.5.3
numpy==1.21.6
tqdm==4.66.2
scikit-learn==1.0.2
PyYAML==6.0.1
scipy==1.7.3
matplotlib==3.5.3
plotly==5.18.0
chardet==5.2.0




